add some tests

+++ microblogs/models.py
@@ -30,7 +30,7 @@ class Post(models.Model):
     )
     text = models.CharField(max_length = 280)
     created_at = models.DateTimeField(auto_now_add=True)
-    id = models.BigAutoField(primary_key=True)
+    id = models.IntegerField(primary_key=True)
 
     class Meta:
         ordering = ['-created_at']
